Decomposition
Deck feature decomposition and some code refactoring
