+++ SetGame/SetGame/Features/Settings.swift
@@ -9,14 +9,21 @@ import Foundation
 import ComposableArchitecture
 
 struct Settings: Reducer {
+    
+    // MARK: -
+    // MARK: State
     struct State: Equatable {
         var isShowHint: Bool = true
     }
     
+    // MARK: -
+    // MARK: Action
     enum Action: Equatable {
         case showHintToggle
     }
     
+    // MARK: -
+    // MARK: Reducer
     var body: some ReducerOf<Self> {
         Reduce { state, action in
             switch action {
