+++ SetGame/SetGame/Features/Statistic.swift
@@ -9,32 +9,41 @@ import Foundation
 import ComposableArchitecture
 
 struct Statistic: Reducer {
-    struct GameStatistic: Equatable {
-        var secondsElapsed: UInt = 0
-        var foundSets: UInt = 0
-    }
-    
     @Dependency(\.continuousClock) private var clock
+    
+    // MARK: -
+    // MARK: State
     struct State: Equatable {
         var currentGame = GameStatistic()
         var bestGame = GameStatistic()
+        
+        struct GameStatistic: Equatable {
+            var secondsElapsed: UInt = 0
+            var foundSets: UInt = 0
+            var usedHints: UInt = 0
+        }
     }
     
+    // MARK: -
+    // MARK: State
     enum Action: Equatable {
         case begin
         case end
         case tick
         case reset
-        case foundSet
+        case setFound
+        case useHint
     }
 
     private struct TimerID: Hashable {}
     
+    // MARK: -
+    // MARK: Reducer
     var body: some ReducerOf<Self> {
         Reduce { state, action in
             switch action {
             case .begin:
-                state.currentGame = GameStatistic()
+                state.currentGame = State.GameStatistic()
                 return .run { send in
                     for await _ in clock.timer(interval: .seconds(1)) {
                         await send(.tick)
@@ -59,19 +68,23 @@ struct Statistic: Reducer {
                 return .none
                 
             case .reset:
-                state.currentGame = GameStatistic()
+                state.currentGame = State.GameStatistic()
                 
                 return .send(.end)
                 
-            case .foundSet:
+            case .setFound:
                 state.currentGame.foundSets += 1
+                return .none
+                
+            case .useHint:
+                state.currentGame.usedHints += 1
                 return .none
             }
         }
     }
 }
 
-extension Statistic.GameStatistic {
+extension Statistic.State.GameStatistic {
     var timeElapsed: String {
         var calendar = Calendar.current
         let formatter = DateComponentsFormatter()
