+++ SetGame/SetGame/Views/Shapes/StripedBackground.swift
